// database.types.new.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4";
  };
  public: {
    Tables: {
      users: {
        Row: {
          user_id: string;
          email: string;
          created_at: string;
          updated_at: string | null;
          is_active: boolean;
          metadata: Json;
        };
        Insert: {
          user_id?: string;
          email: string;
          created_at?: string;
          updated_at?: string | null;
          is_active?: boolean;
          metadata?: Json;
        };
        Update: {
          user_id?: string;
          email?: string;
          created_at?: string;
          updated_at?: string | null;
          is_active?: boolean;
          metadata?: Json;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};